chore(rpc): centralize taikoAuth batch lookups and error handling
